refactor(ui): move robot configuration into a single file
This file should contain all logic needed to ocnfigure specific robots
for the UI.
Later we may look into moving this into the backend so that there will
be only 1 place where we need to configure this.
