+++ application/ui/src/features/robots/setup-wizard/shared/setup-robot-viewer.tsx
@@ -10,8 +10,9 @@ import { degToRad } from 'three/src/math/MathUtils.js';
 import { URDFRobot } from 'urdf-loader';
 
 import { useContainerSize } from '../../../../components/zoom/use-container-size';
-import { urdfPathForType, useLoadModelMutation, useRobotModels } from '../../robot-models-context';
+import { useLoadModelMutation, useRobotModels } from '../../robot-models-context';
 import { SchemaRobotType } from '../../robot-types';
+import { urdfPathForType } from '../../robots-configuration';
 import { JointHighlight, useJointHighlight } from './use-joint-highlight';
 
 // ---------------------------------------------------------------------------
